added others

+++ lib/screen/Home/notes/note_detail_page.dart
@@ -1,7 +1,11 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
+import 'package:notesync/shared/toast.dart';
+import '../../../services/auth.dart';
 import '../add_notes.dart';
 import 'package:notesync/database/firestore.dart';
+
+import '../shared_methods.dart';
 
 class NoteDetailPage extends StatelessWidget {
   final String date;
@@ -17,6 +21,8 @@ class NoteDetailPage extends StatelessWidget {
   Widget build(BuildContext context) {
     Color primaryColor = const Color.fromRGBO(33, 133, 176, 1);
     FirestoreService firestoreService = FirestoreService();
+
+    final String? currentUserId = AuthService().currentUser?.uid;
 
     // Fetch the note data once as a stream
     Stream<DocumentSnapshot> noteStream =
@@ -40,37 +46,50 @@ class NoteDetailPage extends StatelessWidget {
         var noteData = snapshot.data!;
         String title = noteData['title'] ?? 'Untitled';
         String body = noteData['body'] ?? 'No content available';
+        String ownerId =
+            noteData['userId']; // Fetch the ownerId from the note document
 
         return Scaffold(
           appBar: AppBar(
             title: const Text("Note Details"),
             backgroundColor: primaryColor,
             actions: [
-              // Edit Icon
-              IconButton(
-                icon: const Icon(Icons.edit),
-                onPressed: () {
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                      builder: (context) => AddNotes(
-                        title: title,
-                        body: body,
-                        id: id, // Pass the note's id
+              if (currentUserId == ownerId) ...[
+                IconButton(
+                  icon: const Icon(Icons.edit),
+                  onPressed: () {
+                    Navigator.push(
+                      context,
+                      MaterialPageRoute(
+                        builder: (context) => AddNotes(
+                          title: title,
+                          body: body,
+                          id: id, // Pass the note's id
+                        ),
                       ),
-                    ),
-                  );
-                },
-              ),
-              // Delete Icon
-              IconButton(
-                icon: const Icon(Icons.delete),
-                onPressed: () async {
-                  // Delete the note
-                  await firestoreService.deleteNote(id);
-                  Navigator.pop(context); // Return to the previous screen
-                },
-              ),
+                    );
+                  },
+                ),
+                // Delete Icon
+                IconButton(
+                  icon: const Icon(Icons.delete),
+                  onPressed: () async {
+                    showAlert(
+                      context: context,
+                      title: "Delete Note",
+                      content:
+                          "Are you sure you want to delete this note? This action cannot be undone.",
+                      onConfirm: () {
+                        // Delete logic here
+                        firestoreService.deleteNote(id);
+                        showToast(message: "Deleted Succesfully");
+                        Navigator.pop(context); // Return to the previous screen
+                      },
+                      confirmText: "Delete",
+                    );
+                  },
+                ),
+              ]
             ],
           ),
           body: SingleChildScrollView(
